refactored into a new PendingPCDIntegration file
