+++ packages/passport-interface/index.ts
@@ -3,6 +3,7 @@ export * from "./src/Participant";
 export * from "./src/PassportInterface";
 export * from "./src/PassportPopup";
 export * from "./src/PCDIntegration";
+export * from "./src/PendingPCDIntegration";
 export * from "./src/PendingPCDUtils";
 export * from "./src/RequestTypes";
 export * from "./src/SemaphoreGroupIntegration";
